separated validation functions for authorization.
